putchar update 1

+++ 0x02-functions_nested_loops/0-putchar.c
@@ -5,7 +5,7 @@
 
 int main(void)
 {
-	int count;
+	long unsigned int count;
 	char string[] = "_putchar";
 
 	for (count = 0; count < sizeof(string)-1; count++)
